chain: Improve import statements
The import statements througout the code base are already pretty clean,
we can do some minor improvements after observing a couple of things:

- `core`/`alloc` imports are non-uniform, sometimes at the top and
sometimes mixed in with the `crate` imports for `io` and `sync` etc.
While there does seem to be some merit in keeping the core related crate
modules (`io`, `sync`) imports together with crate this does make it
unusual for devs coming from another project - does the benefit outweigh
the cost?

- The rust-bitcoin crate re-exports a bunch of crates and these are
typically kept separate from the main bitcoin import. This seems like a
win for clarity.

- We are importing bitcoin types from the module they are in instead of
using the re-exports. Using the re-exports is beneficial for a couple of
reasons: 1. it reduces maintenance cost because we don't have to change
code when upstream moves things. 2. It is more terse and allows for less
liens of code in imports statements at the cost of more merge conflicts
for PRs that touch bitcoin imports.

Improve the import statements in the `chain` module by doing:

- Order and group imports core, external deps, crate, super
- Order and group bitcoin re-exported crates separately
- Use bitcoin re-exports instead of more fully qualified paths